flow: cumulative WINDOW_UPDATE semantics + hierarchical DEBUG matching
BREAKING WIRE CHANGE (redeployed atomically fleet-wide).

The core fix this commit lands is that flow-control grants are now loss-
tolerant on unreliable transports. Until now WINDOW_UPDATE frames were
sent via writeFrame with SeqNo=0 and bypassed every reliability
mechanism — the per-stream retransmit queue, per-stream sendWindow,
packet-level anti-replay — on the assumption that "control frames are
cheap, don't bother retransmitting". On TCP/WS that held: the TCP layer
below retransmits the segment, so a WINDOW_UPDATE never actually goes
missing. On Noise-UDP it did not: one lost UDP datagram permanently
drained the sender's send credit, the next gossip exchange timed out
after ConsumeTimeout, the session died, and the "no WINDOW_UPDATE ever
reached sender" deadlock was the actual root cause of the remaining
gossip credit stalls on anchor nodes.

Fix: the WINDOW_UPDATE payload now carries the peer's CUMULATIVE total
of credit granted since stream start rather than a per-frame delta.
Cumulative semantics are idempotent across loss, duplication, and
reordering — exactly the property we need:

  - A lost WINDOW_UPDATE is implicitly retransmitted by the next grant
    (which carries a still-larger cumulative value).
  - A duplicate has delta == 0 and is dropped by ApplyUpdate.
  - A reordered older frame has delta <= 0 and is dropped.

This is the same design used by HTTP/2 MAX_STREAM_DATA and QUIC's
window advertisements. StreamWindow and ConnWindow both grow two
counters: grantsEmitted on the receiver side (emitted as the wire
value) and grantsReceived on the sender side (the high watermark seen).
ApplyUpdate computes delta = credit - grantsReceived, drops non-positive
deltas, applies positive deltas with the existing cap-at-dataOutstanding
safety. Tests cover idempotency (duplicate apply is a no-op) and loss
recovery (skipping intermediate grants still fully recovers).

Wire format: WINDOW_UPDATE payload is now 8 bytes (uint64) instead of
4 bytes (uint32). Cumulative totals would wrap the old uint32 at 4 GB
after ~40 hours of heavy gossip traffic, making post-wrap grants appear
stale to the sender and deadlocking the stream. uint64 gives headroom
for centuries of operation at any realistic rate. All four sessions
(TCP, WS, Noise, BrowserWS) plus the WindowUpdater callback and
envelope WindowUpdateFrame.Credit field are migrated together.

Secondary fix: DEBUG namespace matching was using substring search in
the wrong direction — DEBUG=aether did not enable aether.flow because
"aether.flow" is not a substring of "aether". Matching is now:

    DEBUG=*                          — everything
    DEBUG=aether                     — aether and all aether.* descendants
    DEBUG=aether.flow                — just aether.flow and descendants
    DEBUG=aether.flow,aether.gossip  — both explicitly

Entries are comma-separated and matched as (ns == entry ||
strings.HasPrefix(ns, entry + ".")), which is the well-known hierarchical
namespace rule. Flow package's internal debug logger uses the same rule
(duplicated because flow cannot import the parent aether package).

+++ adapter/deliver.go
@@ -13,7 +13,7 @@ import (
 
 // WindowUpdater is called with the stream ID and credit amount to send
 // a WINDOW_UPDATE frame to the remote peer.
-type WindowUpdater func(streamID uint64, credit uint32)
+type WindowUpdater func(streamID uint64, credit uint64)
 
 // DeliveryStats tracks per-stream delivery metrics for monitoring and
 // adaptive behavior. Counters are atomic for lock-free access from
@@ -111,11 +111,14 @@ func DeliverToRecvCh(recvCh chan<- []byte, payload []byte, window *flow.StreamWi
 }
 
 // grantCredit calls ReceiverConsume and sends a WINDOW_UPDATE if the
-// auto-grant threshold is met.
+// auto-grant threshold is met. The returned value from ReceiverConsume is
+// the CUMULATIVE total granted since stream start (not a delta) — it is
+// passed as-is to the WINDOW_UPDATE payload so the sender's ApplyUpdate
+// can compute the delta and drop stale/duplicate frames.
 func grantCredit(window *flow.StreamWindow, payload []byte, streamID uint64, sendUpdate WindowUpdater) {
 	if window != nil && sendUpdate != nil {
 		if grant := window.ReceiverConsume(int64(len(payload))); grant > 0 {
-			sendUpdate(streamID, uint32(grant))
+			sendUpdate(streamID, uint64(grant))
 		}
 	}
 }
